api-v1/requirements.txt
Django==5.0.1
django_filter==23.2
django_seed==0.3.1
djangorestframework_simplejwt==5.3.0
drf_social_oauth2==2.1.3
python-dotenv==1.0.1
stripe==7.12.0
celery
psycopg2
django-cors-headers
Pillow
mysqlclient
python-telegram-bot
